+++ src/main/java/gogitek/restaurentorder/service/UserService.java
@@ -7,6 +7,7 @@ public interface UserService {
     boolean registerUser(User user);
     boolean checkExist(String email);
     User getCurrentUser();
-    void updateUser(int id, User userRequest);
+    void updateUser(Long id, User userRequest);
+
     boolean updatePassword(PasswordDTO passwordDTO);
 }
